feat(simple_server): PR
